+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+-   Update `@typescript-eslint` packages to v8
+
 ## [1.0.42] - 2024-08-07
 
 ### Fixed