+++ packages/gov-ui-kit/CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 -   Update `@typescript-eslint` packages to v8
 -   Remove "Summary" section from `ProposalActionTokenMint` module component and update action interface to only support
     one receiver
+-   Update minor and patch NPM dependencies
 
 ## [1.0.42] - 2024-08-07