+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [1.0.43] - 2024-08-13
+
 ### Added
 
 -   Add Aragon logo, remove dark background options and create custom theme using Aragon branding for ODS storybook